requirements.txt
djangorestframework==3.15.2
Markdown==3.6
django-filter==24.3
# transformers @ git+https://github.com/huggingface/transformers.git@48101cf8d127bbf22d751c7df118a6ce357e2e27
transformers==4.44.0
opencv-python==4.10.0.84
pillow==10.4.0
Django~=5.1
torch==2.4.0